worker/requirements.txt
psycopg[binary]==3.2.3
aio-pika==9.5.3
python-dotenv==1.0.1
httpx==0.27.2
